Rotation Suffix Time Format
Added RotationFileSuffixTimeFormat field to Logger

default: 2006-01-02T15-04-05.000

It used for naming the rotated log files. If not set, then default value is used.

Added functions to count the precision of the layout passed and test the validity of the supplied layout.
If the layout supplied is not valid then default is used