feat: make FileMode for newly created files configurable (#59)
This allows users to set the desired file permissions as a Logger
field. If unset (zero value), defaults to 640 as before.

I also made two minor additional changes:

Octal numbers are rewritten to the recommended 0o-prefixed
format (since Go 1.13).

The permissions for appending to an existing file were set to 644
before, which is useless and might confuse readers. I set it to 0,
according to Go conventions.

Co-authored-by: Andreas Kurth <andreas.kurth@ionos.com>